Extract sorted message hooks into separate files
